move goal state management and submission logic to custom hook, add ability to toggle hook to either update or create
